Move old components in their folder
