+++ libs/search-widget/src/old-components/results/Results.scss
