+++ libs/search-widget/src/old-components/row/Row.scss
