+++ libs/search-widget/src/old-components/row/Square.svelte
