+++ libs/search-widget/src/old-components/search-input/SearchInput.spec.ts
