+++ libs/search-widget/src/old-components/suggestions/Suggestions.scss
