+++ libs/search-widget/src/old-components/viewer/Entities.scss
