+++ libs/search-widget/src/old-components/viewer/Entities.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
-  import Expander from '../common/expander/Expander.svelte';
-  import { viewerStore } from '../core/stores/viewer.store';
-  import { _ } from '../core/i18n';
-  import { nucliaState } from '../core/stores/main.store';
+  import Expander from '../../common/expander/Expander.svelte';
+  import { viewerStore } from '../../core/stores/viewer.store';
+  import { _ } from '../../core/i18n';
+  import { nucliaState } from '../../core/stores/main.store';
   import { combineLatest, map, Observable } from 'rxjs';
-  import type { EntityGroup } from '../core/models';
+  import type { EntityGroup } from '../../core/models';
   import { tap } from 'rxjs/operators';
-  import Icon from '../common/icons/Icon.svelte';
+  import Icon from '../../common/icons/Icon.svelte';
 
   const allEntities = nucliaState().entities;
   const resourceEntities: Observable<EntityGroup[]> = viewerStore.resourceEntities;
