+++ libs/search-widget/src/old-components/viewer/Header.scss
