+++ libs/search-widget/src/old-components/viewer/Header.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import type { Resource } from '@nuclia/core';
-  import { getWidgetActions } from '../core/stores/main.store';
-  import { formatDate, getCDN } from '../core/utils';
-  import MimeIcon from '../common/icons/MimeIcon.svelte';
+  import { getWidgetActions } from '../../core/stores/main.store';
+  import { formatDate, getCDN } from '../../core/utils';
+  import MimeIcon from '../../common/icons/MimeIcon.svelte';
   import ActionMenu from './menus/ActionMenu.svelte';
 
   export let resource: Resource;
