+++ libs/search-widget/src/old-components/viewer/InputViewer.scss
