+++ libs/search-widget/src/old-components/viewer/Metadata.scss
