+++ libs/search-widget/src/old-components/viewer/Preview.scss
