+++ libs/search-widget/src/old-components/viewer/Viewer.scss
