+++ libs/search-widget/src/old-components/viewer/menus/ActionMenu.scss
