+++ libs/search-widget/src/old-components/viewer/menus/LabelMenu.scss
