+++ libs/search-widget/src/old-components/viewer/menus/LabelMenu.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
   import { createEventDispatcher } from 'svelte';
   import { map } from 'rxjs';
-  import { getCDN } from '../../core/utils';
-  import { nucliaState } from '../../core/stores/main.store';
-  import { viewerStore } from '../../core/stores/viewer.store';
-  import { clickOutside } from '../../common/actions/actions';
-  import Label from '../../common/label/Label.svelte';
-  import type { ParagraphLabels } from '../../core/models';
-  import { LabelSetKind } from '../../../../sdk-core/src';
+  import { getCDN } from '../../../core/utils';
+  import { nucliaState } from '../../../core/stores/main.store';
+  import { viewerStore } from '../../../core/stores/viewer.store';
+  import { clickOutside } from '../../../common/actions/actions';
+  import Label from '../../../common/label/Label.svelte';
+  import type { ParagraphLabels } from '../../../core/models';
+  import { LabelSetKind } from '@nuclia/core';
 
   const dispatch = createEventDispatcher();
   export let position: { top: number; left: number } | undefined = undefined;
