+++ libs/search-widget/src/old-components/viewer/paragraphs/Paragraph.scss
