+++ libs/search-widget/src/old-components/viewer/paragraphs/ParagraphWithAnnotations.scss
