+++ libs/search-widget/src/old-components/viewer/paragraphs/ParagraphWithIcon.scss
