+++ libs/search-widget/src/old-components/viewer/paragraphs/ParagraphWithMenu.scss
