+++ libs/search-widget/src/old-components/viewer/paragraphs/Paragraphs.scss
