+++ libs/search-widget/src/old-components/viewer/previewers/Pdf.scss
