+++ libs/search-widget/src/old-components/viewer/previewers/Pdf.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import { onMount, onDestroy } from 'svelte';
   import { Subject, switchMap, tap } from 'rxjs';
-  import { getFile } from '../../core/api';
-  import { viewerStore } from '../../core/stores/viewer.store';
+  import { getFile } from '../../../core/api';
+  import { viewerStore } from '../../../core/stores/viewer.store';
   import PdfControls from './PdfControls.svelte';
 
   export let src: string;
