+++ libs/search-widget/src/old-components/viewer/previewers/PdfControls.scss
