+++ libs/search-widget/src/old-components/viewer/previewers/Player.scss
