+++ libs/search-widget/src/old-components/viewer/previewers/Youtube.scss
