+++ libs/search-widget/src/widgets/EmbeddedSearch.svelte
@@ -2,11 +2,11 @@
   import { merge, map } from 'rxjs';
   import Toggle from '../common/toggle/Toggle.svelte';
   import { nucliaState, nucliaStore } from '../core/stores/main.store';
-  import Results from './results/Results.svelte';
-  import SearchInput from './search-input/SearchInput.svelte';
+  import Results from '../old-components/results/Results.svelte';
+  import SearchInput from '../old-components/search-input/SearchInput.svelte';
   import { _ } from '../core/i18n';
   import Modal from '../common/modal/Modal.svelte';
-  import Suggestions from './suggestions/Suggestions.svelte';
+  import Suggestions from '../old-components/suggestions/Suggestions.svelte';
 
   export let placeholder = '';
